handle content from multiple attachments
begin to cope with multiple attachments in an email - so an attached webpage will have body indexed.

Keywords getting deduplicated should evade issues from identical text/plain and text/html parts for html email.

If this is the right place to do this sort of thing, I'll add code for pdf attachment text extraction (where content will be certainly different to the email), and possibly xls/xlsx too depending on how many of these we have in practice.

Mailpile is really quite helpful. Happy to chat further -- @smithsam on twitter.