Make sys.path entry independent of location
This makes it possible to invoke "mp" through a symbolic
link from anywhere.

We use as a reference the location of the actual scripts/mailpile
script, with all symlinks resolved, rather than the name the user
used to invoke the script, which may e.g. be a symbolic link in ~/bin/.

As a bonus (?), the script can now be invoked directly as
scripts/mailpile, rather than through the 'mp' symlink.