Simplify mail source tag policy: always visible, never labels
Rationale:

After adding a mail source, whether it's local or remote, it needs
to be visible in the UI. The sidebar is the obvious and expected
place for this.  Conversely, we don't want to overload new users
with visual clutter, so the newly created tags shouldn't be treated
as labels (shown in search result listings).

Now that tag settings can be easily edited and the sidebar organizer
works and makes tag settings easy to find, users can easily deviate
from the default policy and we don't need to try and be clever and
guess when to do what.