1st draft at sensible behaviour when out of disk space. Fixes #1696
   * Add a sys.minfree_mb variable for configuring our space requirements
   * Under sys.minfree_mb, stop downloading and indexing mail
   * Under sys.minfree_mb/2, make Mailpile completely read-only
   * Add a HealthCheck command, add notifications in the UI

The HealthCheck command is mostly used internally, but can be invoked from
the CLI if the user is curious. It communicates and keeps state in the Event
Log.

This method will probably be expanded later as other "system health" issues
are identified which we want to check for; this could include low RAM or
reminders to the user about making/enabling backups.