Docker: Mount the code + config inside the docker container
Mounting the sourcecode + config into the docker image in order to:
- have file changes immediately be reflected in the running application
- persists config when shutting the docker container down

I could not find an easy way in docker/docker-compose to configure the
permissions for the `/Mailpile` and `/mailpile-data` directories so that
the `mailpile` user inside the docker container has proper read-write
access. To work around the issue I use `usermod` to set the `mailpile` uid
to 1000. The issue is documented inside the Dockerfile.