Include spam, trash and other hidden tags in mailbox views.
If a search includes a raw mailbox, we interpret that as a sign that the
user is interested in the contents of that particular mailbox, whether it
has been tagged as spam/trash or not.

This makes the mailbox tag views more useful for debugging since they will
more accurately reflect reality, but arguably less useful for day-to-day
reading of mail. Time will tell whether users like that or not...