Implement `onbeforeunload`-event handler for attachment uploads
If a user uploads an attachment and triggers a full-page-refresh we
want to notify him about in-progress uploads and the fact that the
full-page-refresh (or unloading the document) will cancel the upload.

The `unbeforeunload`-event can prompt the user with either our custom
or a generic message (depending on the user agents implementation)
wether she either wants to continue with the full-page-refresh or wants
to wait for the page to "finish it works".